requirements.txt
langgraph>=0.6
langchain-openai>=0.2
jinja2>=3.1
black>=24.1
networkx>=3.3
django>=5.0
pytest>=8.0
python-dotenv>=1.0